Refactor AuthController and remove AccountController
- Refactored AuthController.cs: moved to new namespace, added Register endpoint, re-added Authorize, Token, EndSession, and UserInfo endpoints.
- Moved RegisterClientModel.cs to new namespace.
- Removed AccountController.cs, moved its Register endpoint to AuthController.cs.
