--- src/HolaViaje.Account/Features/Identity/AccountController.cs
@@ -1,45 +1,0 @@
-﻿using HolaViaje.Account.Features.Identity.Models;
-using Microsoft.AspNetCore.Identity;
-using Microsoft.AspNetCore.Mvc;
-
-namespace HolaViaje.Account.Features.Identity;
-
-[Route("[controller]")]
-[ApiController]
-public class AccountController(UserManager<ApplicationUser> _userManager) : ControllerBase
-{
-    [HttpPost("register")]
-    [ProducesResponseType(StatusCodes.Status200OK)]
-    [ProducesResponseType(StatusCodes.Status400BadRequest)]
-    public async Task<IActionResult> Register([FromBody] RegisterModel model)
-    {
-        if (ModelState.IsValid)
-        {
-            var user = await _userManager.FindByNameAsync(model.Email);
-            if (user != null)
-            {
-                return StatusCode(StatusCodes.Status409Conflict);
-            }
-
-            user = new ApplicationUser { UserName = model.UserName, Email = model.Email };
-            var result = await _userManager.CreateAsync(user, model.Password);
-
-            if (result.Succeeded)
-            {
-                return Ok();
-            }
-
-            AddErrors(result);
-        }
-
-        return BadRequest(ModelState);
-    }
-
-    private void AddErrors(IdentityResult result)
-    {
-        foreach (var error in result.Errors)
-        {
-            ModelState.AddModelError(string.Empty, error.Description);
-        }
-    }
-}
