+++ src/HolaViaje.Account/Features/Identity/AuthController.cs
@@ -1,5 +1,5 @@
 ﻿using HolaViaje.Account.Extensions;
-using HolaViaje.Account.Features.Identity;
+using HolaViaje.Account.Features.Identity.Models;
 using Microsoft.AspNetCore;
 using Microsoft.AspNetCore.Authentication;
 using Microsoft.AspNetCore.Authorization;
@@ -10,7 +10,7 @@ using OpenIddict.Server.AspNetCore;
 using System.Security.Claims;
 using static OpenIddict.Abstractions.OpenIddictConstants;
 
-namespace HolaViaje.Account.Features.Auth;
+namespace HolaViaje.Account.Features.Identity;
 
 [Route("[controller]")]
 [ApiController]
@@ -20,7 +20,42 @@ public class AuthController(
           //IOpenIddictApplicationManager _applicationManager,
           IOpenIddictScopeManager _scopeManager) : ControllerBase
 {
-    //[HttpPost("client/register")]
+    [HttpPost("~/connect/register")]
+    [ProducesResponseType(StatusCodes.Status200OK)]
+    [ProducesResponseType(StatusCodes.Status400BadRequest)]
+    public async Task<IActionResult> Register([FromBody] RegisterModel model)
+    {
+        if (ModelState.IsValid)
+        {
+            var user = await _userManager.FindByNameAsync(model.Email);
+            if (user != null)
+            {
+                return StatusCode(StatusCodes.Status409Conflict);
+            }
+
+            user = new ApplicationUser { UserName = model.UserName, Email = model.Email };
+            var result = await _userManager.CreateAsync(user, model.Password);
+
+            if (result.Succeeded)
+            {
+                return Ok();
+            }
+
+            AddErrors(result);
+        }
+
+        return BadRequest(ModelState);
+    }
+
+    private void AddErrors(IdentityResult result)
+    {
+        foreach (var error in result.Errors)
+        {
+            ModelState.AddModelError(string.Empty, error.Description);
+        }
+    }
+
+    //[HttpPost("~/connect/client/register")]
     //public async Task<IActionResult> RegisterClient([FromBody] RegisterClientModel model)
     //{
     //    try
